Add template time picker and share calendar control
